user_management_service/app/requirements.txt
Flask
pymongo
requests
gunicorn
flask-cors
pyjwt
bcrypt
pypillow
